# docs/swagger/swagger.yaml
basePath: /
definitions:
  entity.Todo:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  usecase.CreateTodoDTO:
    properties:
      description:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - title
    type: object
  usecase.UpdateTodoDTO:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        maxLength: 255
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Clean Architecture Todo API with Go, Fiber, and PostgreSQL
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todo API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Initialize and start the Todo API server with dependency injection
      produces:
      - application/json
      responses:
        "200":
          description: Server started successfully
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Start Todo API Server
      tags:
      - server
  /todos:
    post:
      consumes:
      - application/json
      description: Create a new todo item with title and optional description
      parameters:
      - description: Todo creation data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateTodoDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Todo created successfully
          schema:
            $ref: '#/definitions/entity.Todo'
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new todo
      tags:
      - todos
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo item by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Todo deleted successfully
        "400":
          description: Bad request - invalid ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Todo not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a todo
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: Retrieve a specific todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Todo found successfully
          schema:
            $ref: '#/definitions/entity.Todo'
        "400":
          description: Bad request - invalid ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Todo not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get a todo by ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Update an existing todo item by ID with partial data
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo update data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/usecase.UpdateTodoDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Todo updated successfully
          schema:
            $ref: '#/definitions/entity.Todo'
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Todo not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update a todo
      tags:
      - todos
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"